feat : 코드 리팩토링 friend 진행중

+++ discord_backend/src/main/java/com/imfreepass/discord/user/api/response/ViewState.java
@@ -1,14 +1,9 @@
 package com.imfreepass.discord.user.api.response;
 
-import java.time.ZonedDateTime;
-
-import com.imfreepass.discord.user.entity.State;
-
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 
 @Getter
 @NoArgsConstructor
@@ -16,7 +11,7 @@ import lombok.Setter;
 @Builder
 public class ViewState {
 
-	private Long stateId;
+	private int stateId;
 	private String stateName;
 	private String stateText;
 }
